+++ src/contracts/wolves_dydx_lender.sol
@@ -3,11 +3,24 @@ pragma experimental ABIEncoderV2;
 
 interface IERC20 {
   function totalSupply() external view returns (uint256);
+
   function balanceOf(address account) external view returns (uint256);
+
   function transfer(address recipient, uint256 amount) external returns (bool);
-  function allowance(address owner, address spender) external view returns (uint256);
+
+  function allowance(address owner, address spender)
+    external
+    view
+    returns (uint256);
+
   function approve(address spender, uint256 amount) external returns (bool);
-  function transferFrom(address sender, address recipient, uint256 amount) external returns (bool);
+
+  function transferFrom(
+    address sender,
+    address recipient,
+    uint256 amount
+  ) external returns (bool);
+
   event Transfer(address indexed from, address indexed to, uint256 value);
   event Approval(address indexed owner, address indexed spender, uint256 value);
 }
@@ -22,8 +35,8 @@ struct Set {
 }
 
 enum ActionType {
-  Deposit,   // supply tokens
-  Withdraw  // borrow tokens
+  Deposit, // supply tokens
+  Withdraw // borrow tokens
 }
 
 enum AssetDenomination {
@@ -53,7 +66,7 @@ struct ActionArgs {
 }
 
 struct Info {
-  address owner;  // The address that owns the account
+  address owner; // The address that owns the account
   uint256 number; // A nonce that allows a single address to control many accounts
 }
 
@@ -63,31 +76,46 @@ struct Wei {
 }
 
 interface DyDx {
-  function getAccountWei(Info calldata account, uint256 marketId) external view returns (Wei memory);
+  function getAccountWei(Info calldata account, uint256 marketId)
+    external
+    view
+    returns (Wei memory);
+
   function operate(Info[] calldata, ActionArgs[] calldata) external;
+
   function getEarningsRate() external view returns (Val memory);
-  function getMarketInterestRate(uint256 marketId) external view returns (Val memory);
-  function getMarketTotalPar(uint256 marketId) external view returns (Set memory);
+
+  function getMarketInterestRate(uint256 marketId)
+    external
+    view
+    returns (Val memory);
+
+  function getMarketTotalPar(uint256 marketId)
+    external
+    view
+    returns (Set memory);
 }
 
 library SafeMath {
   function sub(uint256 a, uint256 b) internal pure returns (uint256) {
-    require(b <= a, "SafeMath: Sub failed");
+    require(b <= a, 'SafeMath: Sub failed');
     uint256 c = a - b;
     return c;
   }
+
   function div(uint256 a, uint256 b) internal pure returns (uint256) {
-    require(b > 0, "SafeMath: div failed");
+    require(b > 0, 'SafeMath: div failed');
     uint256 c = a / b;
     return c;
   }
+
   function mul(uint256 a, uint256 b) internal pure returns (uint256) {
     if (a == 0) {
       return 0;
     }
 
     uint256 c = a * b;
-    require(c / a == b, "SafeMath: multiplication overflow");
+    require(c / a == b, 'SafeMath: multiplication overflow');
 
     return c;
   }
@@ -108,22 +136,31 @@ contract DyDxLender {
   address constant sai = 0xC4375B7De8af5a38a93548eb8453a498222C4fF2;
 
   function getId() external pure returns (bytes32) {
-    return keccak256(abi.encodePacked("DyDxLender"));
+    return keccak256(abi.encodePacked('DyDxLender'));
   }
 
   function approve(address token) external {
-    IERC20(token).approve(dydx, uint(-1));
+    IERC20(token).approve(dydx, uint256(-1));
   }
 
-  function invest(address token, uint256 assetAmount) external returns(uint256) {
-    uint dToken = _token2dToken(token);
-    require(dToken != uint(-1));
+  function invest(address token, uint256 assetAmount)
+    external
+    returns (uint256)
+  {
+    uint256 dToken = _token2dToken(token);
+    require(dToken != uint256(-1));
 
     // mint dToken
     Info[] memory infos = new Info[](1);
     infos[0] = Info(address(this), 0);
 
-    AssetAmount memory amt = AssetAmount(true, AssetDenomination.Wei, AssetReference.Delta, assetAmount);
+    AssetAmount memory amt =
+      AssetAmount(
+        true,
+        AssetDenomination.Wei,
+        AssetReference.Delta,
+        assetAmount
+      );
     ActionArgs memory act;
     act.actionType = ActionType.Deposit;
     act.accountId = 0;
@@ -138,15 +175,24 @@ contract DyDxLender {
     return assetAmount;
   }
 
-  function redeem(address token, uint256 poolAmount) external returns (uint256) {
-    uint dToken = _token2dToken(token);
-    require(dToken != uint(-1));
+  function redeem(address token, uint256 poolAmount)
+    external
+    returns (uint256)
+  {
+    uint256 dToken = _token2dToken(token);
+    require(dToken != uint256(-1));
 
     // redeem tokens to this contract
     Info[] memory infos = new Info[](1);
     infos[0] = Info(address(this), 0);
 
-    AssetAmount memory amt = AssetAmount(false, AssetDenomination.Wei, AssetReference.Delta, poolAmount);
+    AssetAmount memory amt =
+      AssetAmount(
+        false,
+        AssetDenomination.Wei,
+        AssetReference.Delta,
+        poolAmount
+      );
     ActionArgs memory act;
     act.actionType = ActionType.Withdraw;
     act.accountId = 0;
@@ -161,36 +207,44 @@ contract DyDxLender {
     return poolAmount;
   }
 
-  function balanceOf(address token, address _owner) external view returns (uint256) {
-    Wei memory bal = DyDx(dydx).getAccountWei(Info(_owner, 0), _token2dToken(token));
+  function balanceOf(address token, address _owner)
+    external
+    view
+    returns (uint256)
+  {
+    Wei memory bal =
+      DyDx(dydx).getAccountWei(Info(_owner, 0), _token2dToken(token));
     return bal.value;
   }
 
   // return the amount of the underlying asset
-  function getAssetAmount(address token, address _owner) external view returns (uint256) {
-    Wei memory bal = DyDx(dydx).getAccountWei(Info(_owner, 0), _token2dToken(token));
+  function getAssetAmount(address token, address _owner)
+    external
+    view
+    returns (uint256)
+  {
+    Wei memory bal =
+      DyDx(dydx).getAccountWei(Info(_owner, 0), _token2dToken(token));
     return bal.value;
   }
 
   function getApr(address token) external view returns (uint256) {
-    uint dToken = _token2dToken(token);
+    uint256 dToken = _token2dToken(token);
     uint256 rate = DyDx(dydx).getMarketInterestRate(dToken).value;
     uint256 aprBorrow = rate * 31622400;
     Set memory set = DyDx(dydx).getMarketTotalPar(dToken);
     uint256 usage = (set.borrow * 1e18) / set.supply;
-    return (((aprBorrow * usage) / 1e18) * DyDx(dydx).getEarningsRate().value) / 1e18;
+    return
+      (((aprBorrow * usage) / 1e18) * DyDx(dydx).getEarningsRate().value) /
+      1e18;
   }
 
-  function refresh(address token) external {
-  }
+  function refresh(address token) external {}
 
-  function _token2dToken(address asset) internal pure returns (uint){
-    if (asset == sai)
-      return 1;
-    if (asset == usdc)
-      return 2;
-    if (asset == dai)
-      return 3;
-    return uint(-1);
+  function _token2dToken(address asset) internal pure returns (uint256) {
+    if (asset == sai) return 1;
+    if (asset == usdc) return 2;
+    if (asset == dai) return 3;
+    return uint256(-1);
   }
 }
